backend/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coverage==7.2.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
httpie==3.2.1
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
mysqlclient==2.1.1
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
PySocks==1.7.1
pytz==2023.3
requests==2.30.0
requests-toolbelt==1.0.0
rich==13.3.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
